Merge r1922246 from trunk:
Don't use AuthFormLoginRequiredLocation in inline

Intro to inline says:

If a non-authenticated user attempts to access a page protected by
mod_auth_form that isn't configured with a AuthFormLoginRequiredLocation
directive, a HTTP_UNAUTHORIZED status code is returned to the browser
indicating to the user that they are not authorized to view the page.

The entire point seems to be to keep the URL the same by using an internal redirect 
via ErrorDocument, and AuthFormLoginRequiredLocation conflicts with it.

Submitted By: Rishikeshan Lavakumar/Sulochana <oss AT @rishikeshan.com> 




git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1922247 13f79535-47bb-0310-9956-ffa450edef68